requirements.txt
#
####### APIPlug Python /Django API requirements.txt #######
#
###### Requirements ######
Django >= 1.9
mysqlclient
djangorestframework
django-rest-swagger
#